Don't do useless and non-portable pointer-to-integer cast
git-svn-id: svn+ssh://svn.cprover.org/var/lib/cvs/satabs/trunk@240 c82bf971-7879-4458-aa57-88e1f0fec4b8

+++ src/satabs/modelchecker/abstract_counterexample.h
@@ -40,7 +40,7 @@ extern inline bool operator<(
     abstract_counterexamplet::stepst::const_iterator a,
     abstract_counterexamplet::stepst::const_iterator b)
 {
-  return (unsigned long int)&(*a)<(unsigned long int)&(*b);
+  return &(*a)<&(*b);
 }
 
 std::ostream& operator<< (std::ostream& out,
